+++ packages/expo-updates/ios/Tests/EXUpdatesFileDownloaderTests.m
@@ -30,7 +30,7 @@
   }];
   EXUpdatesFileDownloader *downloader = [[EXUpdatesFileDownloader alloc] initWithUpdatesConfig:config];
 
-  NSURLRequest *actual = [downloader createManifestRequestWithURL:[NSURL URLWithString:@"https://exp.host/@test/test"]];
+  NSURLRequest *actual = [downloader createManifestRequestWithURL:[NSURL URLWithString:@"https://exp.host/@test/test"] extraHeaders:nil];
   XCTAssertEqual(NSURLRequestReloadIgnoringCacheData, actual.cachePolicy);
   // this fails, seems like this header isn't actually set on the object until later
   // XCTAssertEqualObjects(@"no-cache", [actual valueForHTTPHeaderField:@"Cache-Control"]);
@@ -39,15 +39,57 @@
 - (void)testCacheControl_NewManifest
 {
   EXUpdatesConfig *config = [EXUpdatesConfig configWithDictionary:@{
-    @"EXUpdatesURL": @"https://exp.host/@test/test",
+    @"EXUpdatesURL": @"https://exp.host/manifest/00000000-0000-0000-0000-000000000000",
     @"EXUpdatesRuntimeVersion": @"1.0",
     @"EXUpdatesUsesLegacyManifest": @(NO)
   }];
   EXUpdatesFileDownloader *downloader = [[EXUpdatesFileDownloader alloc] initWithUpdatesConfig:config];
 
-  NSURLRequest *actual = [downloader createManifestRequestWithURL:[NSURL URLWithString:@"https://exp.host/@test/test"]];
+  NSURLRequest *actual = [downloader createManifestRequestWithURL:[NSURL URLWithString:@"https://exp.host/manifest/00000000-0000-0000-0000-000000000000"] extraHeaders:nil];
   XCTAssertEqual(NSURLRequestUseProtocolCachePolicy, actual.cachePolicy);
   XCTAssertNil([actual valueForHTTPHeaderField:@"Cache-Control"]);
+}
+
+- (void)testExtraHeaders_ObjectTypes
+{
+  EXUpdatesConfig *config = [EXUpdatesConfig configWithDictionary:@{
+    @"EXUpdatesURL": @"https://exp.host/manifest/00000000-0000-0000-0000-000000000000",
+    @"EXUpdatesRuntimeVersion": @"1.0"
+  }];
+  EXUpdatesFileDownloader *downloader = [[EXUpdatesFileDownloader alloc] initWithUpdatesConfig:config];
+
+  NSDictionary *extraHeaders = @{
+    @"expo-string": @"test",
+    @"expo-number": @(47.5),
+    @"expo-boolean": @YES
+  };
+
+  NSURLRequest *actual = [downloader createManifestRequestWithURL:[NSURL URLWithString:@"https://exp.host/manifest/00000000-0000-0000-0000-000000000000"] extraHeaders:extraHeaders];
+  XCTAssertEqualObjects(@"test", [actual valueForHTTPHeaderField:@"expo-string"]);
+  XCTAssertEqualObjects(@"47.5", [actual valueForHTTPHeaderField:@"expo-number"]);
+  XCTAssertEqualObjects(@"true", [actual valueForHTTPHeaderField:@"expo-boolean"]);
+}
+
+- (void)testExtraHeaders_OverrideOrder
+{
+  EXUpdatesConfig *config = [EXUpdatesConfig configWithDictionary:@{
+    @"EXUpdatesURL": @"https://exp.host/manifest/00000000-0000-0000-0000-000000000000",
+    @"EXUpdatesRuntimeVersion": @"1.0",
+    @"EXUpdatesRequestHeaders": @{
+      // custom headers configured at build-time should be able to override preset headers
+      @"expo-updates-environment": @"custom"
+    }
+  }];
+  EXUpdatesFileDownloader *downloader = [[EXUpdatesFileDownloader alloc] initWithUpdatesConfig:config];
+
+  // serverDefinedHeaders should not be able to override preset headers
+  NSDictionary *extraHeaders = @{
+    @"expo-platform": @"android"
+  };
+
+  NSURLRequest *actual = [downloader createManifestRequestWithURL:[NSURL URLWithString:@"https://exp.host/manifest/00000000-0000-0000-0000-000000000000"] extraHeaders:extraHeaders];
+  XCTAssertEqualObjects(@"ios", [actual valueForHTTPHeaderField:@"expo-platform"]);
+  XCTAssertEqualObjects(@"custom", [actual valueForHTTPHeaderField:@"expo-updates-environment"]);
 }
 
 @end